Add support for floating point operations.

+++ include/math.h
@@ -1,8 +1,8 @@
 #ifndef __PLATFORM_MATH_HEADER__
 #define __PLATFORM_MATH_HEADER__
 
-int max(int a, int b);
-int min(int a, int b);
+double max(double a, double b);
+double min(double a, double b);
 
 int power(int base, int exp);
